Seems Object#send has the better performance then Object#public_send.

### Environment
- Kubuntu 21.10
- AMD Ryzen 7 5700G
- gcc version 11.2.0
- Ruby 3.1.1

#### Benchmark result
```
Warming up --------------------------------------
         Object#send     1.443M i/100ms
  Object#public_send     1.239M i/100ms
Calculating -------------------------------------
         Object#send     14.514M (± 1.9%) i/s -     73.585M in   5.071831s
  Object#public_send     12.476M (± 0.4%) i/s -     63.214M in   5.066893s

Comparison:
         Object#send: 14514416.9 i/s
  Object#public_send: 12476119.8 i/s - 1.16x  (± 0.00) slower
```

### Test code
```ruby
require 'benchmark/ips'

class Foo
  def foo
    123
  end
end

Benchmark.ips do |x|
  foo = Foo.new

  x.report('Object#send') {
    foo.send(:foo)
  }

  x.report('Object#public_send') {
    foo.public_send(:foo)
  }

  x.compare!
end
```